Moved authentication stuff to pkg
